refactor(test): restructure array tests into success/failure/edge pattern
Replaces 7 per-function submodule files with tests/success.rs,
tests/failure.rs, and tests/edge.rs following the project test layout.
