+++ crates/core/src/eval/functions/array/tests/mod.rs
@@ -1,7 +1,3 @@
-mod rows_columns;
-mod transpose;
-mod sort;
-mod unique;
-mod array_constrain;
-mod sumproduct;
-mod flatten;
+mod success;
+mod failure;
+mod edge;
